requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1

torch~=2.3.0
torchvision~=0.18.0
scikit-learn~=1.5.0